+++ src/test/compile-fail/coherence-all-remote.rs
@@ -13,7 +13,7 @@
 extern crate "coherence-lib" as lib;
 use lib::Remote1;
 
-impl<T> Remote1<T> for int { }
-//~^ ERROR E0117
+impl<T> Remote1<T> for isize { }
+//~^ ERROR E0210
 
 fn main() { }
