+++ src/test/compile-fail/coherence-blanket-conflicts-with-specific-cross-crate.rs
@@ -13,17 +13,17 @@
 extern crate go_trait;
 
 use go_trait::{Go,GoMut};
-use std::fmt::Show;
+use std::fmt::Debug;
 use std::default::Default;
 
 struct MyThingy;
 
 impl Go for MyThingy {
-    fn go(&self, arg: int) { }
+    fn go(&self, arg: isize) { }
 }
 
 impl GoMut for MyThingy { //~ ERROR conflicting implementations
-    fn go_mut(&mut self, arg: int) { }
+    fn go_mut(&mut self, arg: isize) { }
 }
 
 fn main() { }
