+++ src/test/compile-fail/issue-17458.rs
@@ -8,8 +8,8 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-static X: uint = 0 as *const uint as uint;
-//~^ ERROR: can not cast a pointer to an integer in a constant expression
+static X: usize = 0 as *const usize as usize;
+//~^ ERROR: can not cast a pointer to an integer in statics
 
 fn main() {
     assert_eq!(X, 0);
