+++ src/test/compile-fail/issue-17718-const-bad-values.rs
@@ -8,13 +8,13 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-const C1: &'static mut [uint] = &mut [];
-//~^ ERROR: constants are not allowed to have mutable references
+const C1: &'static mut [usize] = &mut [];
+//~^ ERROR: references in constants may only refer to immutable values
 
-static mut S: uint = 3;
-const C2: &'static mut uint = &mut S;
+static mut S: usize = 3;
+const C2: &'static mut usize = &mut S;
 //~^ ERROR: constants cannot refer to other statics
-//~^^ ERROR: are not allowed to have mutable references
+//~^^ ERROR: references in constants may only refer to immutable values
 
 fn main() {}
