+++ src/test/compile-fail/issue-9243.rs
@@ -8,10 +8,10 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-// Regresion test for issue 9243
+// Regression test for issue 9243
 
 struct Test {
-    mem: int,
+    mem: isize,
 }
 
 pub static g_test: Test = Test {mem: 0}; //~ ERROR statics are not allowed to have destructors
