+++ src/test/compile-fail/lint-unused-extern-crate.rs
@@ -12,6 +12,10 @@
 
 #![deny(unused_extern_crates)]
 #![allow(unused_variables)]
+#![allow(deprecated)]
+#![feature(libc)]
+#![feature(collections)]
+#![feature(rand)]
 
 extern crate libc; //~ ERROR: unused extern crate
 
@@ -28,6 +32,6 @@ use rand::isaac::IsaacRng;
 use other::*;
 
 fn main() {
-    let x: collecs::vec::Vec<uint> = Vec::new();
+    let x: collecs::vec::Vec<usize> = Vec::new();
     let y = foo();
 }
