+++ src/test/compile-fail/move-fragments-2.rs
@@ -18,9 +18,11 @@
 // These are checking that enums are tracked; note that their output
 // paths include "downcasts" of the path to a particular enum.
 
+#![feature(rustc_attrs)]
+
 use self::Lonely::{Zero, One, Two};
 
-pub struct D { d: int }
+pub struct D { d: isize }
 impl Drop for D { fn drop(&mut self) { } }
 
 pub enum Lonely<X,Y> { Zero, One(X), Two(X, Y) }
