+++ src/test/compile-fail/move-fragments-3.rs
@@ -18,9 +18,11 @@
 // This checks the handling of `_` within variants, especially when mixed
 // with bindings.
 
+#![feature(rustc_attrs)]
+
 use self::Lonely::{Zero, One, Two};
 
-pub struct D { d: int }
+pub struct D { d: isize }
 impl Drop for D { fn drop(&mut self) { } }
 
 pub enum Lonely<X,Y> { Zero, One(X), Two(X, Y) }
