+++ src/test/compile-fail/moves-based-on-type-block-bad.rs
@@ -10,13 +10,16 @@
 
 // ignore-tidy-linelength
 
+#![feature(box_patterns)]
+#![feature(box_syntax)]
+
 struct S {
     x: Box<E>
 }
 
 enum E {
     Foo(Box<S>),
-    Bar(Box<int>),
+    Bar(Box<isize>),
     Baz
 }
