+++ src/test/compile-fail/privacy5.rs
@@ -9,15 +9,14 @@
 // except according to those terms.
 
 // aux-build:privacy-tuple-struct.rs
-// ignore-fast
 
 extern crate "privacy-tuple-struct" as other;
 
 mod a {
     pub struct A(());
-    pub struct B(int);
-    pub struct C(pub int, int);
-    pub struct D(pub int);
+    pub struct B(isize);
+    pub struct C(pub isize, isize);
+    pub struct D(pub isize);
 
     fn test() {
         let a = A(());
