+++ src/test/compile-fail/static-mut-not-constant.rs
@@ -8,9 +8,10 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
+#![feature(box_syntax)]
 
-static mut a: Box<int> = box 3;
-//~^ ERROR statics are not allowed to have custom pointers
+static mut a: Box<isize> = box 3;
+//~^ ERROR allocations are not allowed in statics
 //~^^ ERROR mutable statics are not allowed to have owned pointers
 
 fn main() {}
