+++ src/test/compile-fail/unboxed-closure-sugar-wrong-trait.rs
@@ -12,8 +12,9 @@
 
 trait Trait {}
 
-fn f<F:Trait(int) -> int>(x: F) {}
-//~^ ERROR wrong number of type arguments: expected 0, found 2
+fn f<F:Trait(isize) -> isize>(x: F) {}
+//~^ ERROR wrong number of type arguments: expected 0, found 1
+//~| ERROR no associated type `Output`
 
 fn main() {}
